requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.0
numpy>=1.19.0
flask>=2.0.0
smbus2>=0.4.1
pyserial>=3.5
pillow>=8.0.0
scikit-image>=0.18.0
pandas>=1.3.0
matplotlib>=3.4.0
requests>=2.26.0
python-dotenv>=0.19.0 